# contracts/openapi/admin-v1.yaml
openapi: 3.0.3
info:
  title: qiuxs-blog Admin API
  version: 1.0.0
servers:
  - url: https://blog-admin.qiuxs.com
paths:
  /api/admin/v1/session:
    post:
      operationId: loginAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: Authenticated administrator
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminView'}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '403': {$ref: '#/components/responses/ProblemResponse'}
        '429': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
    delete:
      operationId: logoutAdmin
      responses:
        '204': {description: Session removed or already absent}
        '403': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
  /api/admin/v1/me:
    get:
      operationId: getCurrentAdmin
      responses:
        '200':
          description: Current administrator
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminView'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
  /api/admin/v1/articles:
    get:
      operationId: listArticles
      parameters:
        - name: state
          in: query
          required: false
          schema: {type: string, enum: [active, trashed], default: active}
      responses:
        '200': {$ref: '#/components/responses/ArticleListResponse'}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
    post:
      operationId: createArticle
      responses:
        '201': {$ref: '#/components/responses/ArticleDetailResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '403': {$ref: '#/components/responses/ProblemResponse'}
        '409': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
  /api/admin/v1/articles/{articleId}:
    parameters:
      - {$ref: '#/components/parameters/ArticleId'}
    get:
      operationId: getArticle
      responses:
        '200': {$ref: '#/components/responses/ArticleDetailResponse'}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '404': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
  /api/admin/v1/articles/{articleId}/draft:
    parameters:
      - {$ref: '#/components/parameters/ArticleId'}
    put:
      operationId: saveArticleDraft
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SaveDraftRequest'}
      responses:
        '200': {$ref: '#/components/responses/DraftViewResponse'}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '403': {$ref: '#/components/responses/ProblemResponse'}
        '404': {$ref: '#/components/responses/ProblemResponse'}
        '409': {$ref: '#/components/responses/ProblemResponse'}
        '422': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
  /api/admin/v1/articles/{articleId}/preview:
    parameters:
      - {$ref: '#/components/parameters/ArticleId'}
    get:
      operationId: getArticlePreview
      responses:
        '200':
          description: Current draft preview
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PreviewView'}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '404': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
  /api/admin/v1/articles/{articleId}/versions:
    parameters:
      - {$ref: '#/components/parameters/ArticleId'}
    get:
      operationId: listArticleVersions
      responses:
        '200':
          description: Frozen article revisions
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RevisionList'}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '404': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
    post:
      operationId: createArticleVersion
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LockVersionRequest'}
      responses:
        '201':
          description: Frozen version and replacement editing draft
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VersionResult'}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '403': {$ref: '#/components/responses/ProblemResponse'}
        '404': {$ref: '#/components/responses/ProblemResponse'}
        '409': {$ref: '#/components/responses/ProblemResponse'}
        '422': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
  /api/admin/v1/articles/{articleId}/versions/{revisionId}/restore:
    parameters:
      - {$ref: '#/components/parameters/ArticleId'}
      - {$ref: '#/components/parameters/RevisionId'}
    post:
      operationId: restoreArticleVersion
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LockVersionRequest'}
      responses:
        '200': {$ref: '#/components/responses/DraftViewResponse'}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '403': {$ref: '#/components/responses/ProblemResponse'}
        '404': {$ref: '#/components/responses/ProblemResponse'}
        '409': {$ref: '#/components/responses/ProblemResponse'}
        '422': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
  /api/admin/v1/articles/{articleId}/trash:
    parameters:
      - {$ref: '#/components/parameters/ArticleId'}
    post:
      operationId: trashArticle
      responses:
        '204': {description: Article moved to trash}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '403': {$ref: '#/components/responses/ProblemResponse'}
        '404': {$ref: '#/components/responses/ProblemResponse'}
        '409': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
  /api/admin/v1/articles/{articleId}/untrash:
    parameters:
      - {$ref: '#/components/parameters/ArticleId'}
    post:
      operationId: untrashArticle
      responses:
        '204': {description: Article restored from trash}
        '400': {$ref: '#/components/responses/ProblemResponse'}
        '401': {$ref: '#/components/responses/ProblemResponse'}
        '403': {$ref: '#/components/responses/ProblemResponse'}
        '404': {$ref: '#/components/responses/ProblemResponse'}
        '409': {$ref: '#/components/responses/ProblemResponse'}
        '503': {$ref: '#/components/responses/ProblemResponse'}
...
components:
  parameters:
    ArticleId:
      name: articleId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
    RevisionId:
      name: revisionId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
  responses:
    ProblemResponse:
      description: Request failed
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    ArticleListResponse:
      description: Articles in descending update order
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ArticleList'}
    ArticleDetailResponse:
      description: Article and current editing draft
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ArticleDetail'}
    DraftViewResponse:
      description: Current editing draft
      content:
        application/json:
          schema: {$ref: '#/components/schemas/DraftView'}
...
  schemas:
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: {type: string, minLength: 1, maxLength: 64}
        password:
          type: string
          minLength: 1
          maxLength: 256
          description: UTF-8 encoded value must be at most 256 bytes.
    AdminView:
      type: object
      additionalProperties: false
      required: [id, username]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        username: {type: string}
    ArticleSummary:
      type: object
      additionalProperties: false
      required: [id, slug, draftRevisionId, publishedRevisionId, state, draftTitle, draftUpdatedAt, createdAt, updatedAt]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        slug: {type: string}
        draftRevisionId: {type: integer, format: int64, minimum: 1}
        publishedRevisionId: {type: integer, format: int64, minimum: 1, nullable: true}
        state: {type: string, enum: [active, trashed]}
        draftTitle: {type: string}
        draftUpdatedAt: {type: string, format: date-time}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    ArticleList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/ArticleSummary'}
    ArticleDetail:
      type: object
      additionalProperties: false
      required: [id, slug, draftRevisionId, publishedRevisionId, state, createdAt, updatedAt, draft]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        slug: {type: string}
        draftRevisionId: {type: integer, format: int64, minimum: 1}
        publishedRevisionId: {type: integer, format: int64, minimum: 1, nullable: true}
        state: {type: string, enum: [active, trashed]}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        draft: {$ref: '#/components/schemas/DraftView'}
    DraftView:
      type: object
      additionalProperties: false
      required: [id, articleId, revisionNo, lockVersion, status, reason, title, summary, coverMediaId, contentMd, contentHash, tags, media, createdAt, updatedAt]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        articleId: {type: integer, format: int64, minimum: 1}
        revisionNo: {type: integer, format: int64, minimum: 1}
        lockVersion: {type: integer, format: int64, minimum: 1}
        status: {type: string, enum: [editing]}
        reason: {type: string, enum: [draft]}
        title: {type: string}
        summary: {type: string}
        coverMediaId: {type: integer, format: int64, minimum: 1, nullable: true}
        contentMd: {type: string}
        contentHash: {type: string}
        tags:
          type: array
          maxItems: 32
          items: {$ref: '#/components/schemas/TagSnapshot'}
        media:
          type: array
          maxItems: 257
          items: {$ref: '#/components/schemas/MediaReference'}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    PreviewView:
      type: object
      additionalProperties: false
      required: [draft]
      properties:
        draft: {$ref: '#/components/schemas/DraftView'}
    RevisionView:
      type: object
      additionalProperties: false
      required: [id, articleId, revisionNo, lockVersion, status, reason, title, summary, coverMediaId, contentMd, contentHash, tags, media, createdAt, updatedAt]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        articleId: {type: integer, format: int64, minimum: 1}
        revisionNo: {type: integer, format: int64, minimum: 1}
        lockVersion: {type: integer, format: int64, minimum: 1}
        status: {type: string, enum: [frozen]}
        reason: {type: string, enum: [manual_version, publish_snapshot]}
        title: {type: string}
        summary: {type: string}
        coverMediaId: {type: integer, format: int64, minimum: 1, nullable: true}
        contentMd: {type: string}
        contentHash: {type: string}
        tags:
          type: array
          maxItems: 32
          items: {$ref: '#/components/schemas/TagSnapshot'}
        media:
          type: array
          maxItems: 257
          items: {$ref: '#/components/schemas/MediaReference'}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    RevisionList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/RevisionView'}
    VersionResult:
      type: object
      additionalProperties: false
      required: [version, draft]
      properties:
        version: {$ref: '#/components/schemas/RevisionView'}
        draft: {$ref: '#/components/schemas/DraftView'}
    TagSnapshot:
      type: object
      additionalProperties: false
      required: [tagId, name, slug, position]
      properties:
        tagId: {type: integer, format: int64, minimum: 1}
        name: {type: string}
        slug: {type: string}
        position: {type: integer, minimum: 0}
    MediaReference:
      type: object
      additionalProperties: false
      required: [mediaId, publicKey, purpose, position]
      properties:
        mediaId: {type: integer, format: int64, minimum: 1}
        publicKey: {type: string}
        purpose: {type: string, enum: [cover, content]}
        position: {type: integer, minimum: 0}
    SaveDraftRequest:
      type: object
      additionalProperties: false
      required: [lockVersion, title, summary, coverMediaId, contentMd, tagIds]
      properties:
        lockVersion: {type: integer, format: int64, minimum: 1}
        title: {type: string}
        summary: {type: string}
        coverMediaId: {type: integer, format: int64, minimum: 1, nullable: true}
        contentMd: {type: string, maxLength: 2097152}
        tagIds:
          type: array
          maxItems: 32
          uniqueItems: true
          items: {type: integer, format: int64, minimum: 1}
    LockVersionRequest:
      type: object
      additionalProperties: false
      required: [lockVersion]
      properties:
        lockVersion: {type: integer, format: int64, minimum: 1}
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        code: {type: string}
        requestId: {type: string}